Update of addButton method and addition of Type
I've updated the addButton method to read in the element button information from a file so that we don't have to hard code everything in the program. I've also added a Type enum for the classification of each metal. In case you're not super familiar with reading files in java, the elementList.txt file must be placed in the main project folder for execution, not the package folder. I'm working on adjusting the element information in a similar way as well, but let me know what you think of these changes.